fix: rename button component
